requirements.txt
flask==3.1.1
flask-cors==6.0.1
openai>=1.98.0
python-dotenv==1.1.1
pypdf2==3.0.1
python-docx==1.2.0
werkzeug>=3.0.4